TEMP Map Collision
Partially functional map collision, doesn't work on the left or right sides of the screen, but works on the edges of the canvas itself (and thus, the tops and bottom of the screen). Uploading for Mr. Walters to be able to see and to save progress so far. Map data in main file to avoid CORS errors, will address once collision actually works.